Split controller into auth and user.
