+++ src/controllers/auth.js
@@ -1,0 +1,126 @@
+import UserManager from '../components/userManager';
+import Utils from "../components/utils";
+import AuthManager from "../components/authManager"
+import Db from "../components/db"
+
+class AuthController {
+
+    async generateAuthJwt(req, res) {
+        const did = req.body.did;
+        const contextName = req.body.contextName;
+        const authJwt = AuthManager.generateAuthJwt(did, contextName)
+
+        return res.status(200).send({
+            status: "success",
+            authJwt
+        });
+    }
+    
+    /**
+     * Authenticate the user with a signed consent message to obtain a refresh token.
+     * 
+     * Will automatically create the user if they don't already exist.
+     * 
+     * Returns a refresh token, access token and database hostname for connections
+     * 
+     * @param {*} req 
+     * @param {*} res 
+     * @returns 
+     */
+    async authenticate(req, res) {
+        const authJwt = req.body.authJwt;
+        const did = req.body.did;
+        const contextName = req.body.contextName;
+        const signature = req.body.signature;
+        const deviceId = req.body.deviceId;
+
+        // Verify we have a valid signed auth request
+        const isValid = AuthManager.verifyAuthRequest(authJwt, did, contextName, signature)
+        if (!isValid) {
+            return res.status(400).send({
+                status: "fail",
+                data: {
+                    "auth": "Invalid credentials or auth token"
+                }
+            });
+        }
+
+        // Verify the user exists
+        const username = Utils.generateUsername(did, contextName);
+        const user = await UserManager.getByUsername(username);
+
+        // Create the user if they don't exist
+        if (!user) {
+            const response = await UserManager.create(username, signature);
+            if (!response || !response._id) {
+                return res.status(400).send({
+                    status: "fail",
+                    data: {
+                        "auth": "User does not exist and unable to create"
+                    }
+                })
+            }
+        }
+
+        // Generate refresh token
+        const refreshToken = await AuthManager.generateRefreshToken(did, contextName, deviceId)
+        const accessToken = await AuthManager.generateAccessToken(refreshToken, contextName)
+
+        return res.status(200).send({
+            status: "success",
+            refreshToken,
+            accessToken,
+            host: Db.buildHost()
+        });
+    }
+
+    async connect(req, res) {
+        const refreshToken = req.body.refreshToken;
+        const contextName = req.body.contextName;
+
+        const accessToken = await AuthManager.generateAccessToken(refreshToken, contextName);
+
+        if (accessToken) {
+            return res.status(200).send({
+                status: "success",
+                accessToken,
+                host: Db.buildHost()    // required to know the CouchDB host
+                // username: removed, don't think it is needed
+            });
+        }
+        else {
+            return res.status(400).send({
+                status: "fail",
+                data: {
+                    "did": "Invalid refresh token or DID"
+                }
+            });
+        }
+    }
+
+    async regenerateRefreshToken(req, res) {
+        const refreshToken = req.body.refreshToken;
+        const contextName = req.body.contextName;
+
+        const newRefreshToken = await AuthManager.regenerateRefreshToken(refreshToken, contextName);
+
+        if (newRefreshToken) {
+            return res.status(200).send({
+                status: "success",
+                refreshToken: newRefreshToken
+            });
+        }
+        else {
+            return res.status(400).send({
+                status: "fail",
+                data: {
+                    "did": "Invalid refresh token or context name"
+                }
+            });
+        }
+    }
+
+}
+
+const authController = new AuthController();
+export default authController;
